Add tests

+++ arty/unicode.c
@@ -2,5 +2,6 @@
 
 bool arty_is_valid_unicode_codepoint(arty_codepoint_t codepoint)
 {
-    return codepoint < 0x110000 && (codepoint - 0xd800 > 0x07ff);
+    return (unsigned int) codepoint < 0x110000
+        && (((unsigned int) codepoint) - 0xd800) > 0x07ff;
 }
